Smooth moving by capturcus

+++ Vec2D.h
@@ -3,10 +3,10 @@
 
 class Vec2D {
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 	Vec2D();
-	Vec2D(int x, int y);
+	Vec2D(float x, float y);
 	
 	Vec2D operator+(Vec2D other);
 	Vec2D operator%(Vec2D other);
